Removed the atol parameter of assert2NormClose.
Controlling the atol parameter is misleading because of the way
numpy.isclose deals with atol. There is a note on atol in the
documentation of numpy.isclose that says that the choosen atol is suited
for values close to 1 but may be unsuited for other values.
Here the function compare with 0, so for the moment I fixed atol=1e-7
and I will see if the value is OK or if I need to change it again.