2 new types for type hints: SUdMatrix and GenericArray.
The SUdMatrix type is a generalisation of the SU2Matrix type. It allows
to show that the concerned parameter is not restricted to a 2x2 matrix
and that the function should also work for SU(d) matrices with d an
integer.
The GenericArray type is a generalisation of the GenericMatrix type. A
GenericMatrix has always 2 dimensions, a GenericArray can have any number
of dimensions.