New data structure: QuantumGateSequence.
A QuantumGateSequence is composed of a set of basis quantum gates and a
sequence of indices pointing to a gate in the basis.
In this implementation, the QuantumGateSequence instances are also
caching the matrix represented by the sequence of basis gates and the
inverses of each basis gates.
This class makes a lot of copies and will probably be refactored or at
least improved in a near future if execution time is a concern.