New data structure: QuantumGateLazyGenerator.
This is a temporary fix and the data structure will be changed by a set of
functions in a near future because there is no need for a class for this kind
of task.
The sequences returned by the data structures may be wrong, tests need to be
performed.
The code is ugly, with repetitions. It is an issue that will be addressed
with the rework explained above.