Implementation of the group_commutator algorithm for SU(2).
The algorithm implemented is the one described by C. Dawson and M. Nielsen
in their article https://arxiv.org/abs/quant-ph/0505030. It was also
implemented in the GitHub of C. Dawson:
    - https://github.com/cmdawson/sk/blob/master/src/su2.cpp#L25
    - https://github.com/cmdawson/sk/blob/master/src/su2.cpp#L111
and in:
    - https://github.com/cryptogoth/skc-python/blob/master/skc/dawson/factor.py#L137
    - https://github.com/cryptogoth/skc-python/blob/master/skc/dawson/factor.py#L185
The major drawback with these implementations was the lack of comments
to explain the code. This issue is addressed in this code.
Tests are missing, I need to figure out the input space of these methods
because they don't seem to work for arbitrary matrices, but still they
give good results within the Solovay-Kitaev algorithm.