Implementation of the similarity_matrix algorithm for SU(2).
The algorithm implemented is the one implemented in the GitHub of C. Dawson:
    - https://github.com/cmdawson/sk/blob/master/src/su2.cpp#L77
and in:
    - https://github.com/cryptogoth/skc-python/blob/master/skc/dawson/factor.py#L10
The major drawback with these implementations was the lack of comments
to explain the code. This issue is not addressed in this code. I still
need to understand the algorithm used.
Tests are missing, I need to figure out the input space of this method
because it does not seem to work for arbitrary matrices, but still it
gives good results within the Solovay-Kitaev algorithm.