SimplificationRule is now a base class and not generic anymore.
Before, the SimplificationRule was generic enough to work for every
rule provided by the user. The problem with this approach is that we are
not restricted to one-qubit operators anymore and the QuantumCircuit
representation changed. Keep the class generic enough was a challenge,
and I decided to write subclasses adapted to specific rules.