Nearest-neighbour structure update.
The nearest-neighbour structure pushed was not the right one. This
commit introduce a preliminary version of the NN structure. For the
moment, the SK implementation suffers from high errors and I don't know
if an approximate nearest-neighbour search (with annoy) would impact the
overall precision. As the search is already implemented, the NN
structure initialise both ANN (annoy) and NN (scipy.cKDTree) and perform
a NN search.